% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{coef}
\alias{coef}
\alias{coef.linreg}
\title{Function to implement coef method for linreg class}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{x}{object of class linreg}
}
\value{
replicate the coef.lm behaviour
}
\description{
Function to implement coef method for linreg class
}
