% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A linear regression package in R.}
\value{
print() prints out the coefficients and coefficient names,
        plot() plots the following two plots using ggplot2,
        resid() returns the vector of residuals e,
        pred() returns the predicted values of ybar,
        coef() returns the coefficients as a named vector,
        summary() returns a similar printout as printed for lm objects.
}
\description{
A package to handle linear regression models. We will use linear algebra to create the most basic functionality in the R package.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula which the model will process.}

\item{\code{data}}{data.frame which is the input data sample.}

\item{\code{parsedata}}{character variable which we need to parse input data.}

\item{\code{beta}}{matrix result for regressions coefficients.}

\item{\code{ybar}}{matrix result for the fitted values.}

\item{\code{ebar}}{matrix which contains the result of the residuals.}

\item{\code{ddf}}{numeric which contains the degrees of freedom.}

\item{\code{sigma_sq}}{numeric which contains the residual variance.}

\item{\code{varian}}{numeric which is the variance of the sample data.}

\item{\code{stdres}}{matrix is the standardized residuals value.}

\item{\code{var_beta}}{matrix is the variance of the regression coefficients.}

\item{\code{t_beta}}{matrix is the t-values for each coefficient.}

\item{\code{p}}{matrix gives the  p-values for each regression coefficient.}
}}


\examples{
linreg_mod <- linreg$new(Petal.Length~Sepal.Width+Sepal.Length,data=iris)
 linreg_mod$print()
 linreg_mod$pred()
 linreg_mod$resid()
 linreg_mod$coef()
 linreg_mod$summary()
 linreg_mod$plot()
          

}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Linear Regression Model}
}
