% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linreg}
\alias{linreg}
\title{Function to implement multiple linear regression without using lm function of R}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{as formula object}

\item{data}{as data.frame}
}
\value{
an object of class linreg as an S3 class
}
\description{
Function to implement multiple linear regression without using lm function of R
}
\examples{
linreg(formula = Petal.Length ~ Species, data = iris)
}
