% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Function to implement plot methd for linreg class using ggplot}
\usage{
\method{plot}{linreg}(x)
}
\arguments{
\item{x}{object of class linreg}
}
\value{
prints the plots to replicate the plot.lm behaviour
}
\description{
Function to implement plot methd for linreg class using ggplot
}
