% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{pred}
\alias{pred}
\title{Function to implement pred method for linreg class}
\usage{
pred(x)
}
\arguments{
\item{object}{of class linreg}
}
\value{
replicate the pred.lm behaviour
}
\description{
Function to implement pred method for linreg class
}
