% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{print.linreg}
\alias{print.linreg}
\title{Function to implement print method for linreg class}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{object of class linreg}

\item{...}{optional object}
}
\value{
prints text to replicate the print.lm behaviour
}
\description{
Function to implement print method for linreg class
}
\examples{
print(linreg(formula = Petal.Length ~ Species, data = iris))
}
