% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{qrlinreg}
\alias{qrlinreg}
\alias{qrdecompositioncalculation}
\title{Function to implement QR Decomposition calculations}
\usage{
qrdecompositioncalculation(formula, data)
}
\arguments{
\item{formula}{as formula object}

\item{data}{as data.frame}
}
\value{
an object of class qrlinreg as an S3 class
}
\description{
Function to implement QR Decomposition calculations
}
\examples{
linreg(formula = Petal.Length ~ Species, data = iris)
}
