% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{resid}
\alias{resid}
\title{Function to implement resd method for linreg class}
\usage{
resid(x)
}
\arguments{
\item{object}{of class linreg}
}
\value{
replicate the resid.lm behaviour
}
\description{
Function to implement resd method for linreg class
}
