% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{summary}
\alias{summary}
\title{Function to implement summary method for linreg classs}
\usage{
summary(x)
}
\arguments{
\item{x}{object of class linreg}
}
\value{
replicate the summary.lm behaviour
}
\description{
Function to implement summary method for linreg classs
}
