% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{coef}
\alias{coef}
\title{Function to implement coef method for linreg class}
\usage{
coef(x)
}
\arguments{
\item{object}{of class linreg}
}
\value{
replicate the coef.lm behaviour
}
\description{
Function to implement coef method for linreg class
}
