% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linreg}
\alias{linreg}
\title{Making a linear regression function using S3 class}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{as object}

\item{data}{as numeric}
}
\value{
calculatedstatistics
}
\description{
Making a linear regression function using S3 class
}
\examples{
linreg(formula = Petal.Length ~ Species, data = iris)
}
