% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{resid}
\alias{resid}
\alias{resid.linreg}
\title{Function to implement resd method for linreg class}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{object of class linreg}
}
\value{
replicate the resid.lm behaviour
}
\description{
Function to implement resd method for linreg class
}
